requirements.txt
discord.py
beautifulsoup4
python-dateutil
tldextract